Moved all implementations to impl
