+++ kydb/impl/tests/test_dynamodb.py
